+++ src/IotHub/IotHub/Properties/Resources.Designer.cs
@@ -86,6 +86,15 @@ namespace Microsoft.Azure.Commands.Management.IotHub.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Add Configuration.
+        /// </summary>
+        internal static string AddIotHubConfiguration {
+            get {
+                return ResourceManager.GetString("AddIotHubConfiguration", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Add Device.
         /// </summary>
         internal static string AddIotHubDevice {
@@ -145,6 +154,15 @@ namespace Microsoft.Azure.Commands.Management.IotHub.Properties {
         internal static string AddIotHubRoutingEndpoint {
             get {
                 return ResourceManager.GetString("AddIotHubRoutingEndpoint", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Invoke Direct Method on a device.
+        /// </summary>
+        internal static string InvokeIotHubDeviceMethod {
+            get {
+                return ResourceManager.GetString("InvokeIotHubDeviceMethod", resourceCulture);
             }
         }
         
@@ -230,6 +248,15 @@ namespace Microsoft.Azure.Commands.Management.IotHub.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Remove Configuration.
+        /// </summary>
+        internal static string RemoveIotHubConfiguration {
+            get {
+                return ResourceManager.GetString("RemoveIotHubConfiguration", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Remove Device.
         /// </summary>
         internal static string RemoveIotHubDevice {
@@ -293,6 +320,15 @@ namespace Microsoft.Azure.Commands.Management.IotHub.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Set Configuration.
+        /// </summary>
+        internal static string SetIotHubConfiguration {
+            get {
+                return ResourceManager.GetString("SetIotHubConfiguration", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Update Distributed Tracing.
         /// </summary>
         internal static string SetIotHubDistributedTracing {
@@ -352,6 +388,15 @@ namespace Microsoft.Azure.Commands.Management.IotHub.Properties {
         internal static string UpdateIotHubModule {
             get {
                 return ResourceManager.GetString("UpdateIotHubModule", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Update Device Module Twin.
+        /// </summary>
+        internal static string UpdateIotHubModuleTwin {
+            get {
+                return ResourceManager.GetString("UpdateIotHubModuleTwin", resourceCulture);
             }
         }