+++ src/Storage/Storage.Management/StorageAccount/StorageAccountBaseCmdlet.cs
@@ -48,6 +48,9 @@ namespace Microsoft.Azure.Commands.Management.Storage
 
         protected const string StorageUsageNounStr = "AzureRmStorageUsage";
 
+        internal const string StandardGZRS = "Standard_GZRS";
+        internal const string StandardRAGZRS = "Standard_RAGZRS";
+
         protected struct AccountAccessTier
         {
             internal const string Hot = "Hot";